banking-services/account-service/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
pydantic==2.0.2
python-multipart==0.0.6
prometheus-client==0.17.0
python-jose[cryptography]==3.3.0
starlette==0.27.0
requests==2.26.0
pyjwt==2.3.0
sqlalchemy>=1.4.0,<2.0.0
pymysql>=1.0.0
jinja2